[BUILD][VSC] Add java-format plugin to Gradle
In order to support a more convenient way of accessing the formatting
feature of google-java-format independently of any IDE it has been
introduced to the gradle build file of the LSP server.
To this time solely for the LSP server but might be rolled out generally
later to the root build file.